created patch api

+++ backend/controllers/workoutsController.js
@@ -13,6 +13,7 @@ const getWorkouts = async (req, res)=>{
 }
 //GET a workout 
 const getWorkout = async(req, res)=>{
+    //check if the id is valid 
   if(!mongoose.Types.ObjectId.isValid(req.params.id)){
     return res.status(404).json({message: "No records found!"});
   }
@@ -43,22 +44,43 @@ const postWorkout = async(req, res)=>{
 }
 //DELETE a workout 
 const deleteWorkout = async(req, res) =>{
-  //check if the workout exist
+    //check if the id is valid 
+   if(!mongoose.Types.ObjectId.isValid(req.params.id)){
+    return res.status(404).json({message: "No records found!"});
+  } 
+    //check if the workout exist
   try {
    const workout = await workoutModels.findById(req.params.id);
    if(!workout){
      return res.status(404).json({error: "No records found!"});
    }
     await workoutModels.deleteOne({_id: req.params.id});
-    return res.status(200).json({message: "Deleted Data Successfull"})
+    return res.status(200).json({message: "Deleted Data Successfull!"})
   } catch (error) {
     return res.status(400).json({error: error.message});
   }
 }
+//PATCH a workout 
+const patchWorkout = async(req, res)=>{
+    //check if the id is valid 
+    if(!mongoose.Types.ObjectId.isValid(req.params.id)){
+        return res.status(404).json({message: "No records found!"});
+    }
+    try {
+        const workout = await workoutModels.findOneAndUpdate({_id: req.params.id}, {...req.body});
+        if(!workout){
+            return res.status(404).json({message: "No records found!"});
+        }
+        return res.status(200).json(workout);
+    } catch (error) {
+       return res.status(400).json({error: error.message}) 
+    }
+} 
 
 module.exports = {
   getWorkouts,
   getWorkout,
   postWorkout,
-  deleteWorkout
+  deleteWorkout,
+  patchWorkout
 }
